requirements.txt
Django==6.0.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-tastypie==0.15.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
